Disable eslint and stylelint
It was a bit tedious to scroll all the way up or down to see some
important information so I decided to comment them out for a while.

Also I added a flag to the package.json build script so that we can
see the output of the bundled modules.